Include libcxxabi in build for Darwin
Previously the libcxx build for Darwin targets (i.e., for the mac host
compiler) did not include libcxxabi, and instead used the one on the
target. But this sometimes has issues with different libcxxabi versions.
Since we already have the mechanism to bake libcxxabi into the build,
use it for Darwin as well to avoid them.

R=jfb@chromium.org, jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3930

Review URL: https://codereview.chromium.org/508033005