Use OSX 10.6 SDK rules for libcxx host build
Our buildbots (unlike my local machine) are using the 10.6 SDK for our host
build of Clang. So switch our localmods to use the appropriate compile
flags and symbol exports for that SDK (we still include libcxxabi in the build
and so do not re-export its symbols).

It also removes the explicit addition of -fPIC to the compile
flags; it's unnecessary (CMake handles fPIC itself correctly) and having multiple entries in that list
seems to be broken.

R=jfb@chromium.org, jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3930

Review URL: https://codereview.chromium.org/530193003